3: started controller
